Use menpo normalize
