--- docs/source/api/menpofit/feature/centralize.rst
@@ -1,7 +1,0 @@
-.. _menpofit-feature-centralize:
-
-.. currentmodule:: menpofit.feature
-
-centralize
-==========
-.. autofunction:: centralize
