--- docs/source/api/menpofit/feature/index.rst
@@ -1,17 +1,0 @@
-.. _api-feature-index:
-
-:mod:`menpofit.feature`
-=======================
-
-Features
---------
-Common feature-related functions required by deformable models.
-
-.. toctree::
-    :maxdepth: 1
-
-    centralize
-    normalize_norm
-    normalize_std
-    normalize_var
-    probability_map
