--- docs/source/api/menpofit/feature/normalize_norm.rst
@@ -1,7 +1,0 @@
-.. _menpofit-feature-normalize_norm:
-
-.. currentmodule:: menpofit.feature
-
-normalize_norm
-==============
-.. autofunction:: normalize_norm
