--- docs/source/api/menpofit/feature/normalize_std.rst
@@ -1,7 +1,0 @@
-.. _menpofit-feature-normalize_std:
-
-.. currentmodule:: menpofit.feature
-
-normalize_std
-=============
-.. autofunction:: normalize_std
