--- docs/source/api/menpofit/feature/normalize_var.rst
@@ -1,7 +1,0 @@
-.. _menpofit-feature-normalize_var:
-
-.. currentmodule:: menpofit.feature
-
-normalize_var
-=============
-.. autofunction:: normalize_var
