--- docs/source/api/menpofit/feature/probability_map.rst
@@ -1,7 +1,0 @@
-.. _menpofit-feature-probability_map:
-
-.. currentmodule:: menpofit.feature
-
-probability_map
-===============
-.. autofunction:: probability_map
