--- menpofit/feature/__init__.py
@@ -1,2 +1,0 @@
-from .features import (
-    centralize, normalize_norm, normalize_std, normalize_var, probability_map)
